Add Qt updater with in-sidebar update bubble; Qt-based release pipeline
- Port UpdateService to Qt (QNetworkAccessManager): GitHub Releases check,
  ETag/304 handling, local cache, dismiss, download progress, silent install
- Inject version via SCREENI_VERSION, override with -DSCREENI_VERSION=x.y.z
- Rewrite installer (Screeni.iss) for Qt exe (Screeni.exe, no WinAppSDK)
- Rewrite CI (release.yml, ci-release.ps1, publish.ps1) for CMake+windeployqt
- Update bubble contained in sidebar, shown only when an update is available
- Embed Monocraft fonts; ignore Qt install dir and artifacts/

+++ src/ui/BarChart.cpp
@@ -21,7 +21,7 @@ void BarChart::setBars(const QVector<ChartBarData>& bars)
 
 QSize BarChart::sizeHint() const
 {
-    const int n = std::max(1, bars_.size());
+    const int n = std::max(1, static_cast<int>(bars_.size()));
     const int w = n * (barWidth_ + 4) + axisWidth_ + 8;
     return QSize(w, plotHeight_ + 40);
 }
